+++ pydov/__init__.py
@@ -5,7 +5,7 @@ import pydov.util.caching
 from pydov.util.hooks import SimpleStatusHook
 
 __author__ = """DOV-Vlaanderen"""
-__version__ = '0.1.3'
+__version__ = '0.2.1'
 
 cache = pydov.util.caching.GzipTextFileCache()
 
@@ -17,4 +17,4 @@ hooks = [
 # session object allows connection pooling and TCP connection reuse.
 request_timeout = 60
 session = requests.Session()
-session.headers.update({'user-agent': 'pydov/%s' % pydov.__version__})
+session.headers.update({'user-agent': 'pydov/{}'.format(pydov.__version__)})