+++ tests/test_types.py
@@ -3,18 +3,29 @@
 import pytest
 
 from pydov.types.boring import Boring
+from pydov.types.fields import XmlField
 from pydov.types.grondwaterfilter import GrondwaterFilter
-from pydov.types.interpretaties import InformeleStratigrafie
-from pydov.search.interpretaties import HydrogeologischeStratigrafie
-from pydov.search.interpretaties import GecodeerdeLithologie
-from pydov.search.interpretaties import LithologischeBeschrijvingen
+from pydov.types.interpretaties import (
+    GecodeerdeLithologie,
+    HydrogeologischeStratigrafie,
+    InformeleStratigrafie,
+    LithologischeBeschrijvingen,
+    FormeleStratigrafie,
+    GeotechnischeCodering,
+    QuartairStratigrafie,
+)
+from pydov.types.sondering import Sondering
 
 type_objects = [Boring,
+                Sondering,
                 GrondwaterFilter,
                 InformeleStratigrafie,
+                FormeleStratigrafie,
                 HydrogeologischeStratigrafie,
                 GecodeerdeLithologie,
-                LithologischeBeschrijvingen,]
+                LithologischeBeschrijvingen,
+                GeotechnischeCodering,
+                QuartairStratigrafie,]
 
 
 @pytest.mark.parametrize("objecttype", type_objects)
@@ -39,3 +50,39 @@ def test_get_fields_sourcexml(objecttype):
     fields = objecttype.get_fields(source=('xml',))
     for field in fields.values():
         assert field['source'] == 'xml'
+
+
+@pytest.mark.parametrize("objecttype", type_objects)
+def test_extend_fields_no_extra(objecttype):
+    """Test the extend_fields method for empty extra_fields.
+
+    Test whether the returned fields match the existing fields.
+    Test whether the returned fields are not the same fields as the original
+    fields.
+
+    """
+    fields = objecttype.extend_fields([])
+    assert fields == objecttype.fields
+    assert fields is not objecttype.fields
+
+
+@pytest.mark.parametrize("objecttype", type_objects)
+def test_extend_fields_with_extra(objecttype):
+    """Test the extend_fields method with extra_fields.
+
+    Test whether the extra field is included.
+
+    """
+    extra_fields = [
+        XmlField(name='grondwatersysteem',
+                 source_xpath='/filter/ligging/grondwatersysteem',
+                 definition='Grondwatersysteem waarin de filter hangt.',
+                 datatype='string',
+                 notnull=False)
+    ]
+
+    fields = objecttype.extend_fields(extra_fields)
+
+    assert len(fields) == len(objecttype.fields) + len(extra_fields)
+
+    assert fields[-1] == extra_fields[-1]